web_app/app/requirements.txt
APScheduler==3.8.0
attrs==21.2.0
backports.entry-points-selectable==1.1.0
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
distlib==0.3.3
filelock==3.3.0
Flask==2.0.2
Flask-APScheduler==1.12.2
freezegun==1.1.0
greenlet==1.1.2
idna==3.3
importlib-metadata==4.8.1
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
numpy==1.21.2
packaging==21.0
pandas==1.3.3
platformdirs==2.4.0
pluggy==1.0.0
psycopg2-binary==2.9.1
py==1.10.0
pyparsing==3.0.1
pytest==6.2.5
pytest-freezegun==0.4.2
python-dateutil==2.8.2
python-dotenv==0.19.1
pytz==2021.3
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.25
toml==0.10.2
typing-extensions==3.10.0.2
tzlocal==2.1
urllib3==1.26.7
virtualenv==20.8.1
Werkzeug==2.0.2
zipp==3.6.0
